Some tweaks to get beframe working as required
It will no longer create a dedicated scratch buffer for new frames.

Additionally, it now requires that the buffer chosen when switching
buffers already exists. This is to prevent accidentally creating a
buffer with a truncated name if the completion is not expanded
before hitting RET. This means if there is only one matching buffer
left, hitting RET will switch to that buffer without first needing
to hit TAB.